+++ app/src/main/java/com/kiero/data/parent/alarm/api/AlarmService.kt
@@ -1,7 +1,7 @@
-package com.kiero.data.alarm.api
+package com.kiero.data.parent.alarm.api
 
 import com.kiero.core.network.model.BaseResponse
-import com.kiero.data.alarm.dto.response.AlarmFeedResponseDto
+import com.kiero.data.parent.alarm.dto.response.AlarmFeedResponseDto
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
